review_sort_2 날짜 데이터 수정 -> 정확한 정렬을 위해 ReviewItem 의 date를 yyyy-MM-dd'T'HH:mm:ss 형식으로 저장, ui에 보여줄 때 마다 parseDate() 메서드로 yyyy-MM-dd 형식으로 변경해서 출력
기존 방법 : ReviewItem 의 date를 yyyy-MM-dd 형식으로 저장 > 정렬할 때 날짜 까지만으로 정렬 가능
