+++ app/src/main/java/com/example/wishhair/review/ReviewItem.java
@@ -93,20 +93,7 @@ public class ReviewItem implements Serializable {
     }
 
     public void setCreatedDate(String createdDate) {
-        this.createdDate = parseDate(createdDate);
-    }
-
-    private String parseDate(String inputDate) {
-        SimpleDateFormat inputFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss", Locale.getDefault());
-        SimpleDateFormat outputFormat = new SimpleDateFormat("yyyy.MM.dd", Locale.getDefault());
-
-        try {
-            Date date = inputFormat.parse(inputDate);
-            return outputFormat.format(date);
-        } catch (ParseException e) {
-            e.printStackTrace();
-        }
-        return "failParseDate";
+        this.createdDate = createdDate;
     }
 
     public String getHairStyleName() {
